Work on Unit Tests
- Split ECB and CBC tests.
- Fix failing ECB test related with exceptions returning from the native code.
- Rename cyphers and test vectors for CBC.
- Fix NIST test vectors for CBC.
- Fix NIST test results for CBC.
- Remove unused code.
- Update README.
