+++ Tests/System.Security.CryptographyTests/System.Security.CryptographyTests.nfproj
@@ -27,7 +27,8 @@
     <RunSettingsFilePath>$(SolutionDir)\.runsettings</RunSettingsFilePath>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="AesTests.cs" />
+    <Compile Include="AesTests_CBC.cs" />
+    <Compile Include="AesTests_ECB.cs" />
     <Compile Include="HmacSha256Tests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
